# templates/oas/model-manager-oas.yaml
openapi: 3.0.3
info:
  title: Model Manager
  description: Model Manager
  version: 1.0.0
servers:
  - url: 'http://model-manager.os4ml.svc.cluster.local:8000'
paths:
...
  "/apis/v1beta1/modelmanager/solutions":
    get:
      summary: 'get all solutions'
      operationId: getSolutions
      parameters:
        - $ref: './shared-oas.yaml#/components/parameters/User'
      tags:
        - modelmanager
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './shared-oas.yaml#/components/schemas/Solution'
    post:
      summary: "Create a Solution"
      operationId: createSolution
      parameters:
        - $ref: './shared-oas.yaml#/components/parameters/User'
      tags:
        - modelmanager
      requestBody:
        content:
          application/json:
            schema:
              $ref: './shared-oas.yaml#/components/schemas/Solution'
      responses:
        201:
          description: "Solution created"
          content:
            application/json:
              schema:
                $ref: './shared-oas.yaml#/components/schemas/Solution'
  "/solutions/{solutionName}":
    get:
      summary: "Get a solution"
      operationId: getSolutionByName
      parameters:
        - $ref: '#/components/parameters/SolutionName'
        - $ref: './shared-oas.yaml#/components/parameters/User'
      tags:
        - modelmanager
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: './shared-oas.yaml#/components/schemas/Solution'
        404:
          $ref: '#/components/responses/NotFound'
    put:
      summary: "Update a Solution"
      operationId: updateSolutionByName
      parameters:
        - $ref: '#/components/parameters/SolutionName'
        - $ref: './shared-oas.yaml#/components/parameters/User'
      tags:
        - modelmanager
      requestBody:
        content:
          application/json:
            schema:
              $ref: './shared-oas.yaml#/components/schemas/Solution'
      responses:
        200:
          description: "Solution updated"
          content:
            application/json:
              schema:
                $ref: './shared-oas.yaml#/components/schemas/Solution'
    delete:
      summary: "Delete a Solution"
      operationId: deleteSolutionByName
      parameters:
        - $ref: '#/components/parameters/SolutionName'
        - $ref: './shared-oas.yaml#/components/parameters/User'
      tags:
        - modelmanager
      responses:
        204:
          $ref: '#/components/responses/Deleted'
...
components:
  schemas:
    Error:
      required:
        - path
        - error
        - timestamp
        - message
        - status
...
  parameters:
...
    SolutionName:
      name: solutionName
      in: path
      description: Name of the Solution
      required: true
      schema:
        type: string
  responses:
...
    Deleted:
      description: The specified resource was deleted
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...